Moved serverPort value into TransientConfig.
