+++ engine/src/main/java/org/terasology/config/NetworkConfig.java
@@ -30,9 +30,6 @@ public class NetworkConfig implements Iterable<ServerInfo> {
     private List<ServerInfo> servers = Lists.newArrayList();
     // Available upstream bandwidth in kilobits per second
     private int upstreamBandwidth = 1024;
-    
-    // the port that is used for hosting
-    private int serverPort = TerasologyConstants.DEFAULT_PORT;
 
     public void clear() {
         servers.clear();
@@ -44,14 +41,6 @@ public class NetworkConfig implements Iterable<ServerInfo> {
 
     public void setUpstreamBandwidth(int upstreamBandwidth) {
         this.upstreamBandwidth = upstreamBandwidth;
-    }
-    
-    public int getServerPort() {
-        return serverPort;
-    }
-
-    public void setServerPort(int serverPort) {
-        this.serverPort = serverPort;
     }
 
     @Override
